+++ src/futhark-test.hs
@@ -168,7 +168,8 @@ runTestCase (TestCase mode program testcase progs) =
         context ("Compiling with " <> T.pack compiler) $ do
           compileTestProgram extra_options compiler program warnings
           mapM_ (testMetrics progs program) structures
-          context "Running compiled program" $
+          unless (mode == Compile) $
+            context "Running compiled program" $
             accErrors_ $ map (runCompiledEntry program progs) ios
       unless (mode == Compile || mode == Compiled) $
         context ("Interpreting with " <> T.pack interpreter) $
@@ -396,7 +397,7 @@ runTests config paths = do
 
   let (excluded, included) = partition (excludedTest config) all_tests
   _ <- forkIO $ mapM_ (putMVar testmvar) included
-  isTTY <- (&& not (configUnbufferOutput config)) <$> hIsTerminalDevice stdout
+  isTTY <- (&& not (configLineOutput config)) <$> hIsTerminalDevice stdout
 
   let report = if isTTY then reportTable else reportText
       clear  = if isTTY then clearFromCursorToScreenEnd else putStr "\n"
@@ -478,7 +479,7 @@ data TestConfig = TestConfig
                   { configTestMode :: TestMode
                   , configPrograms :: ProgConfig
                   , configExclude :: [T.Text]
-                  , configUnbufferOutput :: Bool
+                  , configLineOutput :: Bool
                   }
 
 defaultConfig :: TestConfig
@@ -493,7 +494,7 @@ defaultConfig = TestConfig { configTestMode = Everything
                              , configExtraOptions = []
                              , configExtraCompilerOptions = []
                              }
-                           , configUnbufferOutput = False
+                           , configLineOutput = False
                            }
 
 data ProgConfig = ProgConfig
@@ -555,9 +556,9 @@ commandLineOptions = [
   , Option "C" ["compile"]
     (NoArg $ Right $ \config -> config { configTestMode = Compile })
     "Only compile, do not run."
-  , Option [] ["nobuffer"]
-    (NoArg $ Right $ \config -> config { configUnbufferOutput = True })
-    "Do not buffer output, and write each result on a line by itself."
+  , Option [] ["no-terminal", "notty"]
+    (NoArg $ Right $ \config -> config { configLineOutput = True })
+    "Provide simpler line-based output."
   , Option [] ["typechecker"]
     (ReqArg (Right . changeProgConfig . setTypeChecker) "PROGRAM")
     "What to run for type-checking (defaults to 'futhark')."